requirements.txt
discord.py
youtube_dl
ffmpeg-python
python-dotenv
